Revert version file back to prefixed
We require the version file to be prefixed temporarily until we make changes to the version calculator.

We would have preferred just the version number so ECR is not duplicated in terms of the named tag however due to the current restrictions github will lose its context on the tags. Story raised in the backlog for work to be undertaken.