Add docker-launcher script to image
There was a problem using the container in Concourse in terms of running docker in docker.
Initially we were starting dockerd in the task in Concourse but it was still causing issues.

This commit adds a docker-launcher script to the image which allows us to call a start_docker function.
The script will later be edited to a higher standard and make use of platform-tools-common with regards to logging etc. We wil also look to add the docker aspect to a platform-tools-docker as the script will likely be used again in other projects. A story will be brought into the next sprint to address this.

Resolves: DVOP-1454