Updates Dockerfile and docker-launcher script
A new build of the container will break the container due to a newer version of docker-ce installed which is only applicable for centos-8 and doesn't bring in `iptables` as a dependency. This results in Docker not being able to start up successfully.

Yum installing `iptables` will allow Docker to run successfully.

In addition to this the docker-launcher script has been updated with syntax and whitespacing to follow CH best practice guides.

Plans are still to install docker-launcher via rpm in a story in the current sprint.